minors

+++ src/Gaufrette/Adapter/AzureBlobStorage/BlobProxyFactory.php
@@ -22,7 +22,7 @@ class BlobProxyFactory implements BlobProxyFactoryInterface
      */
     public function __construct($connectionString)
     {
-        if (!class_exists(ServicesBuilder::class)) {
+        if (!class_exists(ServicesBuilder::class) && !class_exists(BlobRestProxy::class)) {
             throw new \LogicException('You need to install package "microsoft/azure-storage-blob" to use this adapter');
         }
         $this->connectionString = $connectionString;
